refractor input and header components
